rename files from jupyter to pyodide
to make it clear that python is being served using WebAssembly and not the jupyter, changing all the references
