+++ app/reducers/pyodideReducer.js
@@ -9,8 +9,8 @@ import {
   SET_PSD_PLOT,
   SET_TOPO_PLOT,
   SET_ERP_PLOT,
-  RECEIVE_EXECUTE_RETURN,
-} from '../epics/jupyterEpics';
+  RECEIVE_EXECUTE_RETURN
+} from '../epics/pyodideEpics';
 import { ActionType, Kernel } from '../constants/interfaces';
 import { KERNEL_STATUS } from '../constants/constants';
 import { EXPERIMENT_CLEANUP } from '../epics/experimentEpics';
